+++ components/src/core/Drawer/Drawer.tsx
@@ -1,39 +1,27 @@
 import React, { useState } from 'react';
-import { createStyles, makeStyles, Theme, useTheme } from '@material-ui/core/styles';
+import { makeStyles, useTheme } from '@material-ui/core/styles';
 import { Drawer, DrawerProps } from '@material-ui/core';
-import Hidden from '@material-ui/core/Hidden';
 import PropTypes from 'prop-types';
 import { DrawerBodyProps } from './DrawerBody';
 import clsx from 'clsx';
 
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        paper: {
-            overflow: 'hidden',
-            position: 'unset',
-        },
-        paperMobile: {
-            maxWidth: '85%',
-            width: theme.spacing(45),
-        },
-        smooth: {
-            transition: 'width 175ms cubic-bezier(.4, 0, .2, 1)',
-        },
-        content: {
-            display: 'flex',
-            flexDirection: 'column',
-            height: '100%',
-           width: '100%'
-        },
-    })
-);
+const useStyles = makeStyles({
+    paper: {
+        overflow: 'hidden',
+        position: 'unset',
+    },
+    content: {
+        display: 'flex',
+        flexDirection: 'column',
+        height: '100%',
+        width: '100%',
+    },
+});
 
 type DrawerClasses = {
     root?: string;
     content?: string;
-    drawer?: string;
     paper?: string;
-    smooth?: string;
 };
 
 // type shared by Drawer, DrawerBody, DrawerNavGroup, NestedNavItem
@@ -134,10 +122,14 @@ export const DrawerComponent: React.FC<DrawerComponentProps> = (props) => {
         ripple,
         titleColor,
         width,
+        variant,
         ...drawerProps // for Material-UI's Drawer component
     } = props;
 
-    const isDrawerOpen = (): boolean => hover || open;
+    const isDrawerOpen = (): boolean => {
+        if (variant === 'persistent') return hover || open;
+        return open;
+    };
 
     const findChildByType = (type: string): JSX.Element[] =>
         React.Children.map(props.children, (child: any) => {
@@ -215,44 +207,28 @@ export const DrawerComponent: React.FC<DrawerComponentProps> = (props) => {
         </>
     );
 
-    const getMobileNavigationMenu = (): JSX.Element => (
-        <Drawer {...props} open={isDrawerOpen()} classes={{ paper: clsx(defaultClasses.paperMobile, props.classes.paper) }}>
-            <div className={clsx(defaultClasses.smooth, classes.smooth, defaultClasses.content, classes.content)}>
+    const defaultContentWidth = theme.spacing(45);
+    const defaultNavigationRailWidth = theme.spacing(7);
+
+    const containerWidth = isDrawerOpen() ? width || defaultContentWidth : defaultNavigationRailWidth;
+    const contentWidth = width || defaultContentWidth;
+
+    return (
+        <Drawer
+            {...drawerProps}
+            variant={variant === 'temporary' ? variant : 'permanent'}
+            open={isDrawerOpen()}
+            classes={{ paper: clsx(defaultClasses.paper, props.classes.paper) }}
+            style={{
+                minHeight: '100%',
+                width: containerWidth,
+                transition: 'width 175ms cubic-bezier(.4, 0, .2, 1)',
+            }}
+        >
+            <div className={clsx(defaultClasses.content, classes.content)} style={{ width: contentWidth }}>
                 {getDrawerContents()}
             </div>
         </Drawer>
-    );
-
-    const getDesktopNavigationMenu = (): JSX.Element => {
-        const containerWidth = isDrawerOpen() ? width || theme.spacing(45) : theme.spacing(7);
-        const contentWidth = width || theme.spacing(45);
-        return (
-            <>
-                <Drawer
-                    {...drawerProps}
-                    variant={'permanent'}
-                    open={isDrawerOpen()}
-                    classes={{ paper: clsx(defaultClasses.paper, props.classes.paper) }}
-                    style={{ minHeight: '100%' }}
-                >
-                    <div
-                        className={clsx(defaultClasses.smooth, classes.smooth)}
-                        style={{ width: containerWidth, height: '100%' }}
-                    >
-                        <div className={clsx(defaultClasses.content, classes.content)} style={{ width: contentWidth }}>
-                            {getDrawerContents()}
-                        </div>
-                    </div>
-                </Drawer>
-            </>
-        );
-    };
-
-    return (
-        <>
-            <Hidden smUp>{getMobileNavigationMenu()}</Hidden>
-            <Hidden xsDown>{getDesktopNavigationMenu()}</Hidden>
-        </>
     );
 };
 
@@ -285,15 +261,10 @@ DrawerComponent.propTypes = {
         root: PropTypes.string,
         content: PropTypes.string,
         paper: PropTypes.string,
-        paperMobile: PropTypes.string,
-        smooth: PropTypes.string,
     }),
     open: PropTypes.bool.isRequired,
     width: PropTypes.number,
     ...PXBlueDrawerNavGroupInheritablePropertiesPropTypes,
-};
-DrawerComponent.defaultProps = {
-    variant: 'permanent',
 };
 DrawerComponent.defaultProps = {
     classes: {},