+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## v3.0.0
-- Adds support for nested items in the Drawer component; adjusted Drawer API.
+- Adds support for nested items in the Drawer component; adjusted Drawer and DrawerLayout API.
 
 **Breaking Changes:** 
 - A few props got renamed to avoid further ambiguities: 
@@ -40,6 +40,7 @@
         ]}
     />
     ```
+- DrawerLayout's `drawer` prop now takes `JSX.Element` instead of `ReactNode`.
 
 ## v2.1.0
 - Adds InfoListTag Component